16.989 s -> toThrowErrorMatchingInlineSnapshot.test.ts (10.366 s) *Promise.*.ts (4.229 s)
